footer and header routes connected
